backend/requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Cache
redis==5.0.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
firebase-admin==6.4.0

# HTTP Client
httpx==0.27.0
aiohttp==3.9.3

# Scheduler
apscheduler==3.10.4

# Data
pydantic==2.6.1
pydantic-settings==2.1.0
python-dateutil==2.8.2

# RSS
feedparser==6.0.11

# Scraping
beautifulsoup4==4.12.3
lxml==5.1.0

# Images
Pillow==10.2.0
opencv-python-headless>=4.10.0

# Avatar Disney (script)
openai>=1.0.0

# Utils
python-dotenv==1.0.1
deep-translator==1.11.4

# WebSocket
websockets==12.0

# Test
pytest>=7.0.0,<8
pytest-asyncio==0.23.4
httpx==0.27.0
